Removed erroneous div and renamed to appropriate file names
