+++ test/features/step_definitions/success-visit-steps.js
@@ -7,7 +7,7 @@ module.exports = function (world) {
 
   world.then('I see text for documents needed for a Class C license', function (done) {
     browser
-      .waitForSelector('.success-visit-form')
+      .waitForSelector('.success-visit-info')
       .then(() => { done(); })
       .catch((err) => {
         throw err;
